% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_lpmf.R
\name{bernoulli_lpmf}
\alias{bernoulli_lpmf}
\title{Probability mass fucntion}
\usage{
bernoulli_lpmf(prob, occ, lp_sum = FALSE)
}
\arguments{
\item{prob}{A value between 0 to 1 of probability of detection}

\item{occ}{A logical value (either 0 or 1) of observed detection}

\item{lp_sum}{Logical. If true, returns the total sum of the
log likelihood}
}
\value{
Numeric. Bernoulli log probability mass function
}
\description{
Probability mass fucntion
}
\examples{
prob <- prob_detec(env = matrix(c(-1, 0, -0.1, 0.9), ncol = 1), mu =  0, sigl =  1, sigr =  0.9, c =  -0.8, pd = 0.9)
bernoulli_lpmf(prob, occ = c(0, 1, 1, 0))
}
