% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio12_df.R
\docType{data}
\name{bio12_df}
\alias{bio12_df}
\title{cmcc_cm bio1 time serires}
\format{
A data frame with time series with 30 years of bio01 data
}
\usage{
bio12_df
}
\description{
DATASET_DESCRIPTION
}
\details{
The data comes from chelsav1 time serires for
cmcc_cm model. We compute bio12 (Anual precipitation) for a 30 years time series.
}
\keyword{datasets}
