% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio1_df.R
\docType{data}
\name{bio1_df}
\alias{bio1_df}
\title{cmcc_cm bio1 time serires}
\format{
A data frame with time series with 30 years of bio01 data
}
\usage{
bio1_df
}
\description{
DATASET_DESCRIPTION
}
\details{
The data comes from chelsav1 time serires for
cmcc_cm model. We compute bio01 (Anual tas average) for a 30 years time series
}
\keyword{datasets}
