% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_envdata_hastime.R
\name{check_envdata_hastime}
\alias{check_envdata_hastime}
\title{Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords}
\usage{
check_envdata_hastime(envData)
}
\arguments{
\item{envData}{An list with environmental data time serires as elements.}
}
\value{
If all the raster time serires are the same, return the crs string. If not
returns FALSE
}
\description{
Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords
}
\examples{
\dontrun{
# Bioclimatic variable time series
data(cmcc_cm_bio1)
data(cmcc_cm_bio12)
data(cmcc_cm_pr)
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio1_ts <- terra::project(bio1_ts, "EPSG:2169")
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio1_ts)
check_sequence_increment(envData)
}
}
