% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pts_duplicity.R
\name{check_pts_duplicity}
\alias{check_pts_duplicity}
\title{Check the duplicity of points. Thecks if there are
at least two or more points in a same cell for a
given raster set.}
\usage{
check_pts_duplicity(occ, envData)
}
\arguments{
\item{occ}{A SpacVector of points geometries. It could had
builded with the vect method for the data.frame signatiure}

\item{envData}{A list of environmental time series to test
the duplicity}
}
\value{
Logical It is true if there is at least
two (or more) points in the same cell
}
\description{
Check the duplicity of points. Thecks if there are
at least two or more points in a same cell for a
given raster set.
}
\examples{
\dontrun{
occ <- xsdm::mus_virtualis
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio1_ts)
check_pts_duplicity(occ, envData)
}
}
