% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pts_haslonglat.R
\name{check_pts_haslonglat}
\alias{check_pts_haslonglat}
\title{Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords}
\usage{
check_pts_haslonglat(occ)
}
\arguments{
\item{occ}{An occurence data.frame.}
}
\value{
Logical. Is TRUE if the data.frame with
occurrence points has the columns longitude and
latitude
}
\description{
Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords
}
\examples{
\dontrun{
pts <- mus_virtualis
check_pts_haslonglat(pts)
}
}
