% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pts_presence_na.R
\name{check_pts_presence_na}
\alias{check_pts_presence_na}
\title{Check points data frame}
\usage{
check_pts_presence_na(occ)
}
\arguments{
\item{occ}{An occurence species data.frame.}
}
\value{
Logical. Is TRUE if the data.frame with
occurrence points has the columns longitude and
latitude
}
\description{
Check points data frame
}
\examples{
\dontrun{
pts <- mus_virtualis
check_pts_presence_na(pts)
}
}
