% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sequence_increment.R
\name{check_sequence_increment}
\alias{check_sequence_increment}
\title{Check sequence. Function to check if a vector is
a sequence with same increment}
\usage{
check_sequence_increment(seq)
}
\arguments{
\item{seq}{A vector to test if is a constant sequence}
}
\value{
If the increment in the sequence is not the same
returns FALSE. Otherwise returns the increment step
}
\description{
Check sequence. Function to check if a vector is
a sequence with same increment
}
