% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcc_cm_pr.R
\docType{data}
\name{cmcc_cm_pr}
\alias{cmcc_cm_pr}
\title{cmcc_cm_pr Monthly precipitation time serires}
\format{
A raster time series with 35 years of monthly precipitation data
}
\usage{
cmcc_cm_pr
}
\description{
DATASET_DESCRIPTION
}
\details{
The data comes from chelsav1 time serires for
cmcc_cm model.
}
\keyword{datasets}
