% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lewontin-cohen.R
\name{.lewontin_cohen_univariate}
\alias{.lewontin_cohen_univariate}
\title{Univariate Lewontin-Cohen XSDM}
\usage{
.lewontin_cohen_univariate(dat, init, chains, ...)
}
\arguments{
\item{dat}{List of data to be passed to \code{rstan::sampling()}.}

\item{init}{Initial values to be passed to \code{rstan::sampling()}.}

\item{chains}{Ingeter, number of chains to run.}

\item{...}{Additional arguments for \code{rstan::sampling()}.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Univariate Lewontin-Cohen XSDM
}
