% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envGridLtsg.R
\name{envGridLtsg}
\alias{envGridLtsg}
\title{Environmental Grid Array. Returns an array of environmental
combinations from a grid with a length out split.
It can be used to get the probability of detection along
a list of environmental time series}
\usage{
envGridLtsg(envData, param.list, lenght = 100)
}
\arguments{
\item{envData}{A list with environmental raster time series}

\item{param.list}{A list of parameters. Could be null for
an unidimensional model (i. e. one environmental variable in the
env list)}

\item{lenght}{Number of grid splits}
}
\value{
A data.table with probability of detection given a grid of environmental range
}
\description{
Environmental Grid Array. Returns an array of environmental
combinations from a grid with a length out split.
It can be used to get the probability of detection along
a list of environmental time series
}
\examples{
bio1 <- terra::unwrap(cmcc_cm_bio1)
bio12 <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio01 = bio1, bio12 = bio12)
pars <- list(
mu = c(3, 400),
sigl = c(0.5, 90),
sigr = c(0.5, 90),
L = matrix(c(1, 0, 0, 1), ncol = 2)
)
envGridLtsg(envData, pars)

}
