% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xsdm_berti_v2.R
\name{fit_xsdm_berti_v2}
\alias{fit_xsdm_berti_v2}
\title{Fit an xsdm model}
\usage{
fit_xsdm_berti_v2(
  xsdm_object,
  fit = NULL,
  recompile = FALSE,
  normalize = FALSE,
  prior_parameters = list(mu_par_1 = NULL, mu_par_2 = NULL, sigl_par = NULL, sigr_par =
    NULL, c_par_1 = NULL, c_par_2 = NULL, L_par = NULL),
  nchains = NULL,
  compile_standalone = FALSE,
  ...
)
}
\arguments{
\item{xsdm_object}{an xsdm object created with constuctor and validated}

\item{fit}{String. Default is TRUE. Calculate parameters either by sampling,
optimization or Laplace methods. Calculate the mle estimation (default) or
Maximum a posteriori estimation (need to pass jacobian = TRUE  argument).
Possible parameters are c("mle", "map", "mle.laplace", "map.laplace")}

\item{recompile}{Default FALSE. If TRUE, recompile the Lewontin-Cohen model
to use features not availiable out of the box. It is needed
to use laplace and optim methods.}

\item{normalize}{Logical, normalize the time series using the mean
and the standard deviation of the presence-absence provided points}

\item{prior_parameters}{list with parameters passed to prior.}

\item{nchains}{Default NULL. The number of mcmc chains passed to cmdstanr sample function.}

\item{compile_standalone}{Default TRUE. It is useful to export
loglik functions}

\item{...}{argument passed to stan model sampling}
}
\value{
a fitted xsdm object
}
\description{
Fit an xsdm model
}
