% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_prob.R
\name{get_log_prob}
\alias{get_log_prob}
\title{Calculate the Log posterior from a fitted xsdm model.}
\usage{
get_log_prob(xsdm_object)
}
\arguments{
\item{xsdm_object}{}
}
\value{
A function that calculates the log of the proabilites
 from a point in the posterior given a fitted xsdm model.
}
\description{
Calculate the Log posterior from a fitted xsdm model.
}
\examples{
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
#Virtual species presence/absence points
pts <- mus_virtualis
if (instantiate::stan_cmdstan_exists()) {
 model <- xsdm(envData, occ = pts, fit = "mle")
 log_prob_f <- get_log_prob(model)
}
}
