% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem.R
\name{idem}
\alias{idem}
\title{Check if two objects are the same}
\usage{
idem(x, y)
}
\arguments{
\item{x}{An R object to test}

\item{y}{An R object to test}
}
\value{
If the objects are the same, returns the object. If
the objects are different returns FALSE
}
\description{
Check if two objects are the same
}
\examples{
\dontrun{
idem("a", "b")
idem("x", "x")
# combined con Reduce can be applied in a list recursively
test <- list("a", "a", "a")
Reduce(idem, test)
#' test <- list("a", "a", "A")
Reduce(idem, test)
}
}
