% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_multivariate.R
\name{lc_multivariate}
\alias{lc_multivariate}
\title{lc_univariate Gets the multivariate Lewontin Cohen model}
\usage{
lc_multivariate()
}
\value{
NULL Is required for it side effects
}
\description{
lc_univariate Gets the multivariate Lewontin Cohen model
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  lc_multivariate()
}
}
\seealso{
Other models: 
\code{\link{lc_univariate}()},
\code{\link{run_lc_model_univariate}()}
}
\concept{models}
