% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_univariate.R
\name{lc_univariate}
\alias{lc_univariate}
\title{lc_univariate Gets the Lewontin Cohen model
univariate}
\usage{
lc_univariate()
}
\value{
NULL Is required for it side effects
}
\description{
lc_univariate Gets the Lewontin Cohen model
univariate
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  lc_univariate()
}
}
\seealso{
Other models: 
\code{\link{run_lc_model_univariate}()}
}
\concept{models}
