% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo}
\alias{loo}
\title{Get loo from xsdm model}
\usage{
loo(xsdm_object, index = 1, ...)
}
\arguments{
\item{xsdm_object}{an xsdm object created with constuctor and validated}

\item{index}{An integer ginving the row index from the sample}

\item{...}{argument passed to stan model sampling}
}
\value{
a fitted xsdm object
}
\description{
Get loo from xsdm model
}
