% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltsg.R
\name{ltsg}
\alias{ltsg}
\title{Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model}
\usage{
ltsg(env, mu, sigl, sigr)
}
\arguments{
\item{env}{Matrix Environmental time series (columns) per sites (rows)}

\item{mu}{numeric parameter.}

\item{sigl}{numeric parameter.}

\item{sigr}{numeric parameter}
}
\value{
A vector with long term stochastic growht values calculated in each site
}
\description{
Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model
}
\examples{
ltsg(env = -1, mu = 0, sigl = 1, sigr = 0.9)
}
