% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltsg_multi.R
\name{ltsg_multi}
\alias{ltsg_multi}
\title{Multivariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model}
\usage{
ltsg_multi(env, mu, sigl, sigr, L)
}
\arguments{
\item{env}{Matrix Environmental time series (columns) per sites (rows)}

\item{mu}{numeric parameter.}

\item{sigl}{numeric parameter.}

\item{sigr}{numeric parameter}

\item{L}{matrix parameter}
}
\description{
Multivariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model
}
