% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltsg_uni.R
\name{ltsg_uni}
\alias{ltsg_uni}
\title{Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model}
\usage{
ltsg_uni(env, mu, sigl, sigr)
}
\arguments{
\item{env}{Matrix Environmental time series (columns) per sites (rows)}

\item{mu}{numeric parameter.}

\item{sigl}{numeric parameter.}

\item{sigr}{numeric parameter}
}
\value{
A vector with long term stochastic growth values calculated in each site
}
\description{
Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model
}
