% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltsg_v2.R
\name{ltsg_v2}
\alias{ltsg_v2}
\title{Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model}
\usage{
ltsg_v2(env, mu, sig)
}
\arguments{
\item{env}{Matrix Environmental time series (columns) per sites (rows)}

\item{mu}{numeric parameter.}

\item{sig}{numeric parameter.}
}
\value{
A vector with long term stochastic growht values calculated in each site
}
\description{
Univariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model
}
\examples{
ltsg_v2(env = -1, mu = 0, sig = 1)
}
