% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xsdm.R
\name{new_xsdm}
\alias{new_xsdm}
\title{Constructor function for the class xsdm}
\usage{
new_xsdm(
  env_data = list(),
  occ = data.frame(),
  stan_model = NULL,
  call = character()
)
}
\arguments{
\item{env_data}{A (prefered named) list with environmental
raster time series.}

\item{occ}{A data.frame}

\item{stan_model}{A stan fitted object slot. Default is NULL}

\item{call}{Call this function}
}
\description{
Constructor function for the class xsdm
}
