% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detec_v2.R
\name{prob_detec_v2}
\alias{prob_detec_v2}
\title{Probability of detection}
\usage{
prob_detec_v2(env, mu, sig, c, pd)
}
\arguments{
\item{env}{Environmental time serires matrix. Rows are sites and
columns are yearly time steps.}

\item{mu}{Numeric parameter. It controls the center of the function}

\item{c}{Numeric parameter. It controls the growth limit.}

\item{pd}{Numeric parameter. It controls the maximum probability}

\item{sigl}{Numeric parameter. It controls the left shape of the bell curve}

\item{sigr}{Numeric parameter. It controls the right shape of the bell curve}

\item{L}{matrix parameter. Default is NULL (i. e. univariate Lewontin Cohen model)}
}
\value{
A numeric value between 0 and 1 with the probability of detection
given the environmental time series matrix and the parameters
}
\description{
Probability of detection
}
\examples{
prob_detec_v2(env = -1, mu =  0, sig =  1, c =  1.1, pd = 0.8)
}
