% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xsdm.R
\name{validate_xsdm}
\alias{validate_xsdm}
\title{validate xsdm_object constructed via xsdm_plus function}
\usage{
validate_xsdm(xsdm_object)
}
\arguments{
\item{xsdm_object}{An xsdm object constructed with new_xsdm function}
}
\value{
A validated xsdm_object
}
\description{
validate xsdm_object constructed via xsdm_plus function
}
