% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_v2.R
\name{vsp_v2}
\alias{vsp_v2}
\title{Virtual species}
\usage{
vsp_v2(env_data, param.list = NULL)
}
\arguments{
\item{env_data}{A list with environmental time series raster object}

\item{param.list}{A list of parameters. Could be null for
an unidimensional model (i. e. one environmental variable in the
env list)}
}
\value{
A raster with the probability of detectio of the
virtual species
}
\description{
Virtual species
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
bio12_ts <- bio12_ts/100
envData <- list(bio1 = bio1_ts, bio12 = bio12_ts)
pars <- list(mu = 4.38, sig = 1.24, c = -8.39, pd = 0.9)
vsp_v2(envData, param.list =  pars)
}
}
