% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnifnot.R
\name{warnifnot}
\alias{warnifnot}
\title{Test the Truth of R Expressions and Warn}
\usage{
warnifnot(..., exprs, exprObject, local = TRUE)
}
\arguments{
\item{..., exprs}{any number of R expressions, which should each evaluate to (a logical vector of all) TRUE.}

\item{exprObject}{alternative to exprs or ...: an ‘expression-like’ object, typically an expression, but also a call, a name, or atomic constant such as TRUE.}

\item{local}{(only when exprs is used:) indicates the environment in which the expressions should be evaluated; by default the one from where stopifnot() has been called.}
}
\description{
Test the Truth of R Expressions and Warn
}
