% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm-package.R
\docType{package}
\name{xsdm-package}
\alias{xsdm-package}
\alias{xsdm}
\title{The 'xsdm' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Emilio Berti \email{emilio.berti.academia@gmail.com} (\href{https://orcid.org/0000-0001-9286-011X}{ORCID})

Other contributors:
\itemize{
  \item Daniel Reuman \email{d294r143@ku.edu} [contributor]
}

}
