% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm.R
\name{xsdm}
\alias{xsdm}
\title{XSDM Species Distribution Modeling}
\usage{
xsdm(x, p, model, ...)
}
\arguments{
\item{x}{List of raster stacks. Each stack represent the timeseries
one climatic variable.}

\item{p}{SpatVector with species' occurrences.}

\item{model}{String, for now only \code{lewontin-cohen}.}

\item{...}{Additional arguments for \code{rstan::sampling()}.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
XSDM Species Distribution Modeling
}
