% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm_berti_v2.R
\name{xsdm_berti_v2}
\alias{xsdm_berti_v2}
\title{Constructor function for the class xsdm}
\usage{
xsdm_berti_v2(env_data = list(), occ = data.frame(), fit = TRUE, ...)
}
\arguments{
\item{env_data}{A (prefered named) list with environmental
raster time series.}

\item{occ}{A data.frame with name of species, longitude, latitude and occurrence column.}

\item{fit}{Logical or string. Default TRUE. Define the fittig
method for xsdm. Default method is sampling from the porsterior
using MCMC. Other possible methods are mle (Maximum likelihood estimation),
map (Maximum a posteriori estimation), mle.laplace (Laplace method),
map.laplace.}

\item{...}{other arguments}
}
\value{
an object of class xsdm
}
\description{
Constructor function for the class xsdm
}
\examples{

bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
#Virtual species presence/absence points
pts <- mus_virtualis
xsdm(envData, occ = pts, fit = FALSE)
if (instantiate::stan_cmdstan_exists()) {
  xsdm(envData, occ = pts, fit = TRUE)
}
}
