% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_raster.R
\name{zero_raster}
\alias{zero_raster}
\title{Title}
\usage{
zero_raster(envData)
}
\arguments{
\item{envData}{env_data A (prefered named) list with environmental
raster time series.}
}
\value{
An empty raster with values of zero given
the environmetal data set provided as list form.
}
\description{
Title
}
\examples{
\dontrun{
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio1_ts)
zero_raster(envData)
}
}
