% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_envdata_spatraster.R
\name{check_envdata_spatraster}
\alias{check_envdata_spatraster}
\title{Check if all elements in env_data list are
Spat Raster}
\usage{
check_envdata_spatraster(envData)
}
\arguments{
\item{envData}{A list of environmental data}
}
\value{
logical A flag if the elements in a env_data list
are not SpatRaster objects from terra package
}
\description{
Check if all elements in env_data list are
Spat Raster
}
