% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_envdata_timestep_consistency.R
\name{check_envdata_timestep_consistency}
\alias{check_envdata_timestep_consistency}
\title{Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords}
\usage{
check_envdata_timestep_consistency(envData)
}
\arguments{
\item{envData}{An list with environmental data time serires as elements.}
}
\value{
If all the raster time serires are the same, return the crs string. If not
returns FALSE
}
\description{
Check environmental data list coordinate reference system
Check if all elements in the env_data list have equal coords
}
