% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xsdm.R
\name{fit_xsdm}
\alias{fit_xsdm}
\title{Fit an xsdm model}
\usage{
fit_xsdm(xsdm_object, ...)
}
\arguments{
\item{xsdm_object}{an xsdm object created with constuctor and validated}

\item{...}{argument passed to stan model sampling}
}
\value{
a fitted xsdm object
}
\description{
Fit an xsdm model
}
