% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xsdm.R
\name{fit_xsdm}
\alias{fit_xsdm}
\title{Fit an xsdm model}
\usage{
fit_xsdm(
  xsdm_object,
  fit = NULL,
  recompile = FALSE,
  compile_standalone = FALSE,
  ...
)
}
\arguments{
\item{xsdm_object}{an xsdm object created with constuctor and validated}

\item{fit}{String. Default is TRUE. Calculate parameters either by sampling,
optimization or Laplace methods. Calculate the mle estimation (default) or
Maximum a posteriori estimation (need to pass jacobian = TRUE  argument).
Possible parameters are c("mle", "map", "mle.laplace", "map.laplace")}

\item{recompile}{Default FALSE. If TRUE, recompile the Lewontin-Cohen model
to use features not availiable out of the box. It is needed
to use laplace and optim methods.}

\item{compile_standalone}{Default TRUE. It is useful to export
loglik functions}

\item{...}{argument passed to stan model sampling}
}
\value{
a fitted xsdm object
}
\description{
Fit an xsdm model
}
