% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltsg_multi.R
\name{ltsg_multi}
\alias{ltsg_multi}
\title{Multivariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model}
\usage{
ltsg_multi(env, mu, sigl, sigr, L)
}
\arguments{
\item{env}{Matrix Environmental time series (columns) per sites (rows)}

\item{mu}{numeric parameter.}

\item{sigl}{numeric parameter.}

\item{sigr}{numeric parameter}

\item{L}{matrix parameter}
}
\value{
A vector with long term stochastic growht values calculated in each site
}
\description{
Multivariate long term stochastic growth function for a
Lewontin Cohen demographic environmental model
}
\examples{
bio1 <- terra::unwrap(xsdm::cmcc_cm_bio1)
bio12 <- terra::unwrap(xsdm::cmcc_cm_bio12)
bio12 <- bio12/100 # rescaled bio12 to be in units with same order of magnitude
envData <- xsdm::envDataArray( xsdm::mus_virtualis, list(bio1, bio12))
L <- matrix(c(1, 0, 0, 1), ncol = 2) #lower triangular matrix
ltsg_multi(env = envData, mu = c(1, 3),   sigl = c(0.5,  0.5), sigr = c(0.5,  0.5),   L = L  )
}
