% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xsdm.R
\name{new_xsdm}
\alias{new_xsdm}
\title{Constructor function for the class xsdm}
\usage{
new_xsdm(env_data = list(), occ = data.frame(), call = character())
}
\arguments{
\item{env_data}{A (prefered named) list with environmental
raster time series.}

\item{occ}{A data.frame}

\item{call}{Call this function}
}
\value{
an object of class xsdm
}
\description{
Constructor function for the class xsdm
}
\examples{
\dontrun{
# example code
# Bioclimatic variable time series
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio1_ts)
#Virtual species presence/absence points
pts <- mus_virtualis
new_xsdm(envData, occ = pts)
}
}
