% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.xsdm}
\alias{predict.xsdm}
\title{Title}
\usage{
\method{predict}{xsdm}(xsdm_object, index = 1, scale = FALSE)
}
\arguments{
\item{xsdm_object}{An xsdm object}

\item{index}{An integer ginving the row index from the sample
of the posterior. Default 1}

\item{scale}{Logical. Should be environmental scaled?
This function takes the mean and standard deviation of provided ocurrence (presence/absence) points.
This scale was used in the fit process.
to weight the environmental data provided to predict}
}
\value{
A raster map
}
\description{
Title
}
\examples{
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
pts <- mus_virtualis
if (instantiate::stan_cmdstan_exists()) {
mod <- xsdm(envData, occ = pts, fit = "map" )
predict(mod)
}
}
