% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\alias{vsp.xsdm}
\title{Virtual species}
\usage{
vsp(xsdm_object, param.list = NULL)

\method{vsp}{xsdm}(xsdm_object, param.list = NULL)
}
\arguments{
\item{xsdm_object}{An xsdm object}

\item{param.list}{A list of parameters}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Virtual species
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
pts <- mus_virtualis
mod <- xsdm(envData, occ = pts, fit = FALSE)
pars <- list(mu = 4.38, sigl = 1.24, sigr = 1.05, c = -8.39, pd = 0.61)
vsp(mod, param.list =  pars) |> terra::plot()
}
}
