% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm-package.R
\docType{package}
\name{xsdm-package}
\alias{xsdm-package}
\title{xsdm: A Demographic Approach to Species Distribution Model}
\description{
Integrates concepts of stochastic demography into species distribution modelling. Inferences, utilizing Bayesian methods based in 'Stan', are made of aspects of the environmental niche of a species, and those inferences can be used to project species geographic range. Data requirements include species detection records and spatiotemporal environmental data in Berti et al. (2024, \doi{10.1101/2024.10.30.621023}) for a description of the statistical and demographic models used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alrobles.github.io/xsdm/}
}

}
\author{
\strong{Maintainer}: Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})

Authors:
\itemize{
  \item Emilio Berti \email{emilio.berti@idiv.de} (\href{https://orcid.org/0000-0001-9286-011X}{ORCID})
  \item Daniel C. Reuman \email{reuman@ku.edu} (\href{https://orcid.org/0000-0002-1407-8947}{ORCID})
}

}
