% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdm.R
\name{xsdm}
\alias{xsdm}
\title{Constructor function for the class xsdm}
\usage{
xsdm(env_data = list(), occ = data.frame(), ...)
}
\arguments{
\item{env_data}{A (prefered named) list with environmental
raster time series.}

\item{occ}{A data.frame}

\item{...}{other arguments}
}
\value{
an object of class xsdm
}
\description{
Constructor function for the class xsdm
}
\examples{

bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
#Virtual species presence/absence points
pts <- mus_virtualis
xsdm(envData, occ = pts)
}
