% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_raster.R
\name{zero_raster}
\alias{zero_raster}
\title{Title}
\usage{
zero_raster(envData)
}
\arguments{
\item{envData}{env_data A (prefered named) list with environmental
raster time series.}
}
\value{
An empty raster with values of zero given
the environmetal data set provided as list form.
}
\description{
Title
}
